tools/requirements.txt
# Python Requirements for Water Change ML Service

# MQTT client
paho-mqtt>=1.6.1

# Machine Learning
numpy>=1.21.0
scikit-learn>=1.0.0

# MariaDB/MySQL Database
mysql-connector-python>=8.0.0

# Environment variable management
python-dotenv>=0.19.0

# Optional: For advanced plotting/analysis
# matplotlib>=3.5.0
# pandas>=1.3.0
